Moving mesh and texture to Enviornment
so that they are accessible when doing ray tests
